Made a start with making entries filter functional
Fixes in the UserSession, now theres a dedicated exception when the user is not logged in, added InvalidOperationException for when the user is not logged in.
Fixes to the validator, it no longer assumes $_POST, now it expects an array of which it filters the field name with.
Added an int sanitizing option
Added a validation for date formatting.