Refactor migrations to exclude ORM usage, resolving errors caused by evolving Entity structures (#37)
Updated the migration approach to avoid using the ORM, as it was leading to issues when Entity structures were modified. Specifically, when the ORM was used to select data from the database, it triggered errors if a column was missing, yet this column was intended to be added in a subsequent migration. This change ensures more robust and error-free migrations.